:rotating_light: Showアクションのテスト作成
元あったテストを分離
